Move loadIfNecessary utils to separate module
